chore: reorganizar carpetas del modelo de login
